--- fisiologia-app/client/src/components/Form/Form.js
@@ -1,82 +1,0 @@
-import React from "react";
-import { useState } from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { postImg } from "../../Redux/Actions/Actions";
-
-function Form() {
-  const dispatch = useDispatch();
-  const [fileInputState, setFileInputState] = useState("");
-  const [previewSource, setpreviewSource] = useState({
-    title: null,
-    img: null,
-    desc: null,
-    podcast: null,
-    grupo: null,
-  });
-
-  const handleInputFile = (e) => {
-    const file = e.target.files[0];
-    previewFile(file);
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    if (!previewSource) return;
-    uploadImage(previewSource);
-  };
-
-  const handleInputText = (e) => {
-    e.preventDefault();
-    setpreviewSource({ ...previewSource, [e.target.name]: e.target.value });
-  };
-
-  const uploadImage = (base64image) => {
-    // console.log(base64image);
-    try {
-      dispatch(postImg(base64image));
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  const previewFile = (file) => {
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-    reader.onloadend = () => {
-      setpreviewSource({ ...previewSource, img: reader.result });
-    };
-  };
-
-  return (
-    <div>
-      <form onSubmit={handleSubmit}>
-        <div className="form">
-          <input
-            type="file"
-            name="image"
-            onChange={handleInputFile}
-            value={fileInputState}
-          />
-        </div>
-        <div className="form">
-          <input type="text" name="title" onChange={handleInputText} />
-        </div>
-        <div className="form">
-          <input type="text" name="grupo" onChange={handleInputText} />
-        </div>
-        <div className="form">
-          <input type="text" name="desc" onChange={handleInputText} />
-        </div>
-        <div className="form">
-          <input type="text" name="podcast" onChange={handleInputText} />
-        </div>
-        <button type="submit">submit</button>
-      </form>
-      {previewSource && (
-        <img src={previewSource.img} alt="img" style={{ height: "300px" }} />
-      )}
-    </div>
-  );
-}
-
-export default Form;
